10line change to float

+++ calculator.py
@@ -7,7 +7,7 @@ def sub(x: int, y: int) -> int:
 def prod(x: int, y: int) -> int:
     return x * y
     
-def div(x: int, y: int) -> int:
+def div(x: int, y: int) -> float:
     return x / y
 
 def hello(txt:str)-> str:
